More Elegant Method to Detect and Remove Byte Order Mark
Sometimes, a text editor may leave a byte order mark (BOM) at the start
of an UTF-8 encoded text file. Such a mark consists of 3 bytes which
define how the file is encoded. This mark is unnecessary if LaTeX is
told to treat the files as UTF-8 anyway. However, if it is present, it
may break the compilation causing LaTeX to die with a strange error
message. This message may be hard to understand for anyone who doesn't
know what UTF-8, let alone what a BOM, is.

A  new script 'removeBOM.sh' has been added which detects the presence
of such a BOM. If it finds it, it will remove it. This means the file
is changed by removing the first 3 bytes. If it does not find a BOM, the
file remains unchanged.

Our TeX compiler script now applies removeBOM.sh to all .tex, .sty, and
.bib files it can find in the directory (recursively).